% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{books}
\alias{books}
\title{Persian Books Dataset}
\format{
A tibble with 105578 rows and 17 variables:
}
\source{
\href{https://github.com/mahmoud-eskandari/PersianBooksDataset}{Mahmud Eskandari}
}
\usage{
books
}
\description{
The dataset contains 105578 books that were published in Iran.
}
\examples{
books

if (require("dplyr")) {
books \%>\%
  count(subject,sort = TRUE) \%>\%
  slice(1:10)
}
}
\keyword{datasets}
