% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadegan_package.R
\docType{package}
\name{dadegan-package}
\alias{dadegan}
\alias{dadegan-package}
\title{dadegan: Some Useful Persian Datasets}
\description{
This package contains a few useful Persian datasets.
    These datasets include factnameh.org's fact-cheking articles, Divar's advertisements, Digikala's purchase hitory and Iranian artists on Spotify.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcnakhaee/dadegan}
  \item Report bugs at \url{https://github.com/mcnakhaee/dadegan/issues}
}

}
\keyword{internal}
