% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spotify}
\alias{spotify}
\title{Spotify}
\format{
A tibble with 10632 rows and 32 variables:
\describe{
  \item{species}{a factor denoting penguin species (Adélie, Chinstrap and Gentoo)}
  \item{island}{a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
  \item{bill_length_mm}{a number denoting bill length (millimeters)}
  \item{bill_depth_mm}{a number denoting bill depth (millimeters)}
  \item{flipper_length_mm}{an integer denoting flipper length (millimeters)}
  \item{body_mass_g}{an integer denoting body mass (grams)}
  \item{sex}{a factor denoting penguin sex (female, male)}
  \item{year}{an integer denoting the study year (2007, 2008, or 2009)}
}
}
\source{
Collected by Muhammad Chenariyan Nakhaee using Spotify's Web API and Spotifyr package)
  \url{mcnakhaee.com}
}
\usage{
spotify
}
\description{
A datas Iranian Artists on
}
\details{
The original data included cases where a country was absent
or was not yet a member. In this dataset these were filtered out to
include only votes of Yes, Abstain, and No.

Country name can be converted to other unique country identifiers
(such as 2-character or 3-character ISO codes) using the
\link[countrycode]{countrycode} function from the countrycode
package.
}
\keyword{datasets}
