% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubaar_train}
\alias{ubaar_train}
\title{Ubaar Transport Cost Prediction (Training)}
\format{
A tibble with 50000 rows and 14 variables:
\describe{
  \item{ID}{an anonymous id unique to a given load }
  \item{date }{the date of the load dispatch. It is a string with ymd format. for example, 960630 is 30th day of 6th month of year 1396.}
  \item{sourceLatitude}{latitude of the source region of transport. Source region is not exactly the location for embarkation of the freight.}
  \item{sourceLongitude }{longitude of the source region of transport. Source region is not exactly the location for embarkation of the freight.}
  \item{SourceState }{state of the source of transport}
  \item{destinationLatitude }{latitude of the destination region of transport. Destination region is not exactly the location for disembarkation of the freight.}
  \item{destinationLongitude }{ latitude of the destination region of transport. Destination region is not exactly the location for disembarkation of the freight.}
  \item{destinationState }{state of the destination of transport}
  \item{distanceKM }{driving distance in kilo meters from source region to destination region for usual cars, calculated using google distance services}
  \item{taxiDurationMin }{estimated arrival time in minutes from source region to destination region for usual cars, calculated using google distance services}
  \item{vehicleType }{ type of the vehicle that has carried the load. It is a categorical variable}
  \item{vehicleOption }{option of the vehicle that has carried the load. It is a categorical variable. Each vehicle type has its own vehicle options}
  \item{weight }{weight of the load that has been carried in Tons (1000KG)}
  \item{price }{transport cost in Rials. This field is the target and you should predict it for test data set.}
}
}
\source{
\href{https://www.kaggle.com/c/ubaar-competition}{Ubaar Kaggle Compettion}
}
\usage{
ubaar_train
}
\description{
In this competition, Ubaar & SADA challenge you to predict transport cost for 15,000 loads. You are provided with data about 50,000 actual transported loads. You can learn more about this competition at http://sanags.ir/ubaar-competition
}
\keyword{datasets}
