% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divar}
\alias{divar}
\title{Divar posts dataset}
\format{
A tibble with 947635 rows and 17 variables:
\describe{
  \item{id}{Unique id of the post (1246772 ids)}
  \item{archive_by_user}{Whether the post was archived by user or automatically by the system (True/False)}
  \item{published_at}{Weekday and hour the post was published }
  \item{cat1}{First level category of the post (for example cat1)}
  \item{cat2}{Second level category of the post (can be empty)}
  \item{cat3}{Third level category of the post (can be empty) (for example light)}
  \item{city}{Name of the city the ad was published in (for example Mashhad)}
  \item{title}{Title of the post. All phone numbers are replaced by the $NUM token.}
  \item{city}{Name of the city the ad was published in (for example Mashhad)}
  \item{desc}{Description of the post. All phone numbers are replaced by the $NUM token.}
  \item{price}{Name of the city the ad was published in (for example Mashhad)}
  \item{image_count}{Number of images for the post}
  \item{platform}{The platform the post was submitted from (mobile or web)}
  \item{mileage}{(Only for light vehicles) the mileage of the vehicle posted in kilometers.}
  \item{brand}{(Only for light vehicles and electronic devices) English and Persian name of the brand separated with ::(two colons)}
  \item{year}{(Only for light vehicles) production year of the vehicle in Iranian calendar. Can be the special value of <1366 which means older than year 1366.}
  \item{type}{(Only for clothing) boys/girls for children clothing and men/women for adult clothing}
}
}
\source{
\href{https://research.cafebazaar.ir}{Cafebazaar Research Group}
}
\usage{
divar
}
\description{
The dataset contains 947635 posts that were published in Divar classified ads platform. These posts were published and archived before 2017. Distribution of these posts in different groups does not resemble the actual distributions.
The dataset contains posts from six major categories in Divar:
\describe{
\item{Vehicles}
\item{Electronic devices}
\item{Businesses}
\item{For the home}
\item{Personal}
\item{Leisure & Hobbies}
}
}
\examples{
divar

if (require("dplyr")) {
divar \%>\%
  select(nutr_id, deriv_id) \%>\%
  left_join(deriv)
}
}
\keyword{datasets}
